# reports/openapi_day4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCM Copilot",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/data/query": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "NL2SQL 查询",
        "description": "自然语言 →（多轮消解）→ SQL（sqlglot 四道闸 + 错误自修复）→ 只读沙箱执行 → 表格 + 洞察。需要权限 data:nl2sql。返回契约见响应模型（含 insights 洞察摘要、repair_log 修复轨迹）。",
        "operationId": "data_query_api_v1_data_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2SqlIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Nl2SqlOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/data/query/{query_id}/feedback": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "SQL 纠错样本回流",
        "description": "SQL 纠错样本回流（feedback 表 fb_type='sql'）——W25 eval_nightly 回流评测集。",
        "operationId": "data_feedback_api_v1_data_query__query_id__feedback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataFeedbackIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataFeedbackOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/scheduler/jobs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "调度任务面板",
        "description": "六任务定义 + 调度状态 + 最近运行历史（job_runs；零重复观测依据）。需要 admin:scheduler:manage。",
        "operationId": "list_scheduler_jobs_api_v1_admin_scheduler_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulerJobsOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/scheduler/jobs/{name}/trigger": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "手动触发调度任务",
        "description": "手动触发任务（写审计；触发的是独立一次性 job，原 cron 不受影响）。需要 admin:scheduler:manage。",
        "operationId": "trigger_scheduler_job_api_v1_admin_scheduler_jobs__name__trigger_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulerTriggerOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataFeedbackIn": {
        "properties": {
          "sql": {
            "type": "string",
            "title": "Sql",
            "description": "实际执行/待纠错的 SQL"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question",
            "description": "对应自然语言问题"
          },
          "correction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction",
            "description": "人工纠正说明"
          },
          "is_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Correct",
            "description": "SQL 是否正确"
          }
        },
        "type": "object",
        "required": [
          "sql"
        ],
        "title": "DataFeedbackIn",
        "description": "SQL 纠错样本回流请求体。"
      },
      "DataFeedbackOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "feedback_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "DataFeedbackOut",
        "description": "反馈回流响应（落库失败返回 ok=false + error，不阻断主链路）。"
      },
      "Err": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "机器可读错误码，如 AUTH_401 / QUOTA_429"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "人类可读错误说明"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "贯穿一次请求的 request_id（排查定位用）"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "Err",
        "description": "统一错误响应体（OpenAPI `responses={401: {\"model\": Err}, ...}` 引用）。"
      },
...
      "JobRunOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "success / skipped / failed"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "执行实例（backend-a1/a2）"
          },
          "trigger": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger",
            "description": "cron / manual"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "JobRunOut",
        "description": "单次任务运行记录（job_runs 一行，零重复观测依据）。"
      },
...
      "Nl2SqlIn": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question",
            "description": "自然语言问题"
          },
          "today": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Today",
            "description": "基准日期（YYYY-MM-DD）；缺省用评测基日，跨月/跨年交给 MySQL"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "多轮会话标识；携带则做指代消解"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "Nl2SqlIn",
        "description": "NL2SQL 查询请求体。\n\n空 question 由服务端显式校验 → 400（模型层不做长度强制，保持既有行为）。"
      },
      "Nl2SqlOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "question": {
            "type": "string",
            "title": "Question",
            "description": "原始问题"
          },
          "resolved_question": {
            "type": "string",
            "title": "Resolved Question",
            "description": "指代消解后的完整问题（无会话 = 原问题）"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "多轮会话标识"
          },
          "table": {
            "type": "boolean",
            "title": "Table",
            "description": "是否返回表格（false = 拒答/降级）"
          },
          "sql": {
            "type": "string",
            "title": "Sql",
            "description": "实际执行的 SQL（100% 透出）"
          },
          "columns": {
            "items": {},
            "type": "array",
            "title": "Columns",
            "description": "列名"
          },
          "rows": {
            "items": {},
            "type": "array",
            "title": "Rows",
            "description": "行数据（已规范化类型）"
          },
          "elapsed": {
            "type": "number",
            "title": "Elapsed",
            "description": "执行耗时 ms"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "description": "是否因行数/字节上限被截断"
          },
          "rejected_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected Reason",
            "description": "被四道闸拒绝的原因"
          },
          "reply": {
            "type": "string",
            "title": "Reply",
            "description": "自然语言回复（摘要/拒答/降级话术）"
          },
          "insights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Insights",
            "description": "结果洞察摘要（≤3 条）"
          },
          "repair_attempts": {
            "type": "integer",
            "title": "Repair Attempts",
            "description": "错误自修复次数（0 = 未修复）",
            "default": 0
          },
          "repair_log": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Repair Log",
            "description": "修复轨迹（generate→error→repair 可回放）"
          }
        },
        "type": "object",
        "required": [
          "question",
          "resolved_question",
          "table",
          "sql",
          "elapsed",
          "truncated",
          "reply"
        ],
        "title": "Nl2SqlOut",
        "description": "NL2SQL 统一响应（表格 + SQL + 洞察；100% 透出 SQL 可审计可纠错）。"
      },
...
      "SchedulerJobOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cron": {
            "type": "string",
            "title": "Cron",
            "description": "cron 表达式"
          },
          "desc": {
            "type": "string",
            "title": "Desc"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "是否已注册到调度器"
          },
          "next_run_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run Time"
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobRunOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/JobRunOut"
            },
            "type": "array",
            "title": "Recent Runs",
            "description": "最近 5 条运行"
          }
        },
        "type": "object",
        "required": [
          "name",
          "cron",
          "desc",
          "enabled"
        ],
        "title": "SchedulerJobOut",
        "description": "调度任务定义 + 调度状态 + 最近运行。"
      },
      "SchedulerJobsOut": {
        "properties": {
          "scheduler": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scheduler",
            "description": "{running, instance, timezone}"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/SchedulerJobOut"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "scheduler",
          "jobs"
        ],
        "title": "SchedulerJobsOut",
        "description": "任务面板响应。"
      },
      "SchedulerTriggerOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "job": {
            "type": "string",
            "title": "Job"
          },
          "triggered": {
            "type": "boolean",
            "title": "Triggered"
          },
          "audited": {
            "type": "boolean",
            "title": "Audited"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "job",
          "triggered",
          "audited"
        ],
        "title": "SchedulerTriggerOut",
        "description": "手动触发响应（触发的是独立一次性 job，不覆盖原 cron）。"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}